% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azuremlweb-package.R
\docType{data}
\name{allsorts}
\alias{allsorts}
\title{A data frame with lots of types in it}
\format{A data frame with 3 rows and 7 variables:
\describe{
  \item{Fac}{A factor variable with levels A,B and C}
  \item{Num}{Floating point number}
  \item{Int}{Strict R integers}
  \item{Intish}{Same as Int but not defined using 1L etc}
  \item{Date}{Standard R date}
  \item{Char}{Character data}
  \item{TimeStamp}{POSIXct stamps}
}}
\source{
Simulated
}
\usage{
allsorts
}
\description{
For use testing whether types are being converted correctly
}
\keyword{datasets}

