% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azuremlweb-package.R
\docType{package}
\name{azuremlweb}
\alias{azuremlweb}
\alias{azuremlweb-package}
\title{azuremlweb: Call an AzureML web service}
\description{
Makes it easier to test out a new AzureML web service. This package helps 
you make a POST request to an AzureML API and parse the response back into R 
data frames.
}

