% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildBody.R
\name{buildBody}
\alias{buildBody}
\title{Build the body of the request}
\usage{
buildBody(datasets)
}
\arguments{
\item{datasets}{A named list of data frames to send to the web service.
The names must match the names of the web inputs in Azure, eg. list(WebInput = data.frame(x=1, y=2))}
}
\value{
A list ready for converting the JSON
}
\description{
Build the body of the request
}
\examples{
buildBody(list(WebInput = data.frame(x=1:5, y=LETTERS[1:5])))
}

