% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildRequestList.R
\name{buildRequestList}
\alias{buildRequestList}
\title{Build request list}
\usage{
buildRequestList(datasets)
}
\arguments{
\item{datasets}{A named list of data frames to send to the web service.
The names must match the names of the web inputs in Azure, eg. list(WebInput = data.frame(x=1, y=2))}
}
\value{
A list ready for converting the JSON
}
\description{
Build request list
}
\examples{
buildRequestList(list(WebInput = data.frame(x=1:5, y=LETTERS[1:5])))
}

