% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildURL.R
\name{buildURL}
\alias{buildURL}
\title{Build an AzureML service URL from workspace info}
\usage{
buildURL(region, workspace, service, details = TRUE)
}
\arguments{
\item{region}{The Azure region, such as ussouthcentral, europewest}

\item{workspace}{The workspace ID}

\item{service}{The service ID}

\item{details}{Logical. Should be TRUE for now}
}
\value{
the url as a character
}
\description{
Build an AzureML service URL from workspace info
}
\examples{
buildURL(region = "ussouthcentral",
         workspace = "8550bbba99572fc47e18e5fdd53e43e2",
         service = "4e7ebd7ad816cd9dd1cb1b17c4cdb0c7",
         details = TRUE)
}

