% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseResponse.R
\name{parseDataFrame}
\alias{parseDataFrame}
\title{Extract one data frame from the}
\usage{
parseDataFrame(result)
}
\arguments{
\item{result}{an element of the response object that corresponds to a single data frame}
}
\value{
A data frame typed according to the ColumnTypes element
}
\description{
Extract one data frame from the
}

