% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseResponse.R
\name{parseResponse}
\alias{parseResponse}
\title{Extract data frames from httr response}
\usage{
parseResponse(resp)
}
\arguments{
\item{resp}{The httr response object}
}
\value{
A list of data frames, one for each output from the web service
}
\description{
Extract data frames from httr response
}

