% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requestService.R
\name{requestService}
\alias{requestService}
\title{Make a POST request to AzureML service}
\usage{
requestService(datasets, url = NULL, region = "ussouthcentral",
  workspace = NULL, service = NULL, auth = NULL, details = TRUE)
}
\arguments{
\item{datasets}{A named list of data frames to send to the web service.
The names must match the names of the web inputs in Azure, eg. list(WebInput = data.frame(x=1, y=2))}

\item{url}{If you supply the full URL then other options are not required}

\item{region}{The Azure region, such as ussouthcentral, europewest}

\item{workspace}{The workspace ID}

\item{service}{The service ID}

\item{auth}{Authentication token. Careful not to save this anywhere and
check your .Rhistory}

\item{details}{Logical. If FALSE then column types are omitted from the
response}
}
\value{
An httr response object
}
\description{
Make a POST request to AzureML service
}
\examples{
\dontrun{
requestService(list(WebInput = data.frame(x=1:5, y=LETTERS[1:5])),
               region = "ussouthcentral",
               workspace = "8550bbba99572fc47e18e5fdd53e43e2",
               service = "4e7ebd7ad816cd9dd1cb1b17c4cdb0c7",
               auth = "TVPYGwNeuZbQLWHjOEKzMWjVihThTcmsIdRRFkSsvHaFoxZcKTjdFmFPaIsfN+OEdoZlGttMJQrzUPOvvLeCej==")
}
}

